SSL.Context: Add support for TLSv1.1 and TLSv1.2; disable SSLv3.
This patch adds support for the TLSv1_1 and TLSv1_2 "methods". It also
documents the "protocol" argument of the Context() constructor,
especially documenting the non-obvious "sslv23" behavior. Last, since
SSLv3 has to be considered insecure at this point, this patch disables
it per default, analogous to the way that SSLv2 is treated.